+++ apps/workflows-dashboard/src/domains/auth/common/types.ts
@@ -5,4 +5,5 @@ export interface IUser {
   lastName: string;
   roles: string[];
   lastActiveAt: Date | null;
+  projectIds?: Array<string>;
 }