+++ packages/common/src/utils/verify-signature/index.ts
@@ -1,0 +1,1 @@
+export { verifySignature } from './verify-signature';