+++ services/workflows-service/scripts/workflows/kyc-email-process-example.ts
@@ -1,4 +1,5 @@
 import { PrismaClient } from '@prisma/client';
+import { env } from '../../src/env';
 
 export const kycEmailSessionDefinition = {
   id: 'kyc_email_session_example',
@@ -74,7 +75,7 @@ export const kycEmailSessionDefinition = {
               transformer: 'jmespath',
               mapping: `{
               endUserId: entity.id,
-              firstName: entity.data.firstName, 
+              firstName: entity.data.firstName,
               lastName: entity.data.lastName,
               callbackUrl: join('',['{secret.APP_API_URL}/api/v1/external/workflows/',workflowRuntimeId,'/hook/KYC_HOOK_RESPONDED', '?resultDestination=pluginsOutput.kyc_session.kyc_session_1.result']),
               vendor: 'veriff'
@@ -114,9 +115,10 @@ export const kycEmailSessionDefinition = {
               receivers: [entity.data.email],
               subject: '{customerCompanyName} activation, Action needed.',
               preheader: 'Verify your identity for Happy Home Goods activation with {customerCompanyName}.',
-              templateId: (documents[].decision[].revisionReason | [0])!=null && 'd-7305991b3e5840f9a14feec767ea7301' || 'd-2c6ae291d9df4f4a8770d6a4e272d803',
+              templateId: (documents[].decision[].revisionReason | [0])!=null && 'd-7305991b3e5840f9a14feec767ea7301' || 'd-61c568cfa5b145b5916ff89790fe2065',
               revisionReason: documents[].decision[].revisionReason | [0],
-              supportEmail: join('',['PayLynk','@support.com']),
+              supportEmail: join('',[entity.data.additionalInfo.companyName,'@support.com'])
+              adapter: '${env.MAIL_ADAPTER}'
               }`, // jmespath
             },
           ],