+++ sdks/workflow-browser-sdk/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ballerine/workflow-browser-sdk
 
+## 0.4.23
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.5.20
+  - @ballerine/workflow-core@0.4.30
+
 ## 0.4.22
 
 ### Patch Changes